the-arena/backend/pyproject.toml
[project]
name = "the-arena-backend"
version = "1.0.0"
description = "The Arena - AI Debate Platform Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "anthropic==0.36.0",
    "chromadb==0.5.5",
    "sentence-transformers==3.0.1",
    "httpx==0.27.2",
    "python-multipart==0.0.9",
    "sse-starlette==2.1.3",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "numpy>=1.26.4",
    "requests>=2.32.3",
    "markdown>=3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.2",
]

[tool.uv.sources]
# Uses default PyPI index
